% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_calendar.R
\name{annotate_calendar_text}
\alias{annotate_calendar_text}
\title{Annotate a Calendar Plot with Text}
\usage{
annotate_calendar_text(
  type = "date",
  colour = "black",
  na_colour = "grey70",
  size = 3,
  digits = 0,
  ...
)
}
\arguments{
\item{type}{What type of annotation should be added? Can be one of "date"
(the day of the month) or "value" (the pollutant concentration).}

\item{colour}{The colour for the annotation text.}

\item{na_colour}{The colour for the annotation text for days outside of a
given month. Used when \code{type = "date"}.}

\item{size}{The size of the annotation text.}

\item{digits}{The number of decimal places used for the annotation text when
\code{type = "value"}.}

\item{...}{Other arguments passed to \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}
}
\description{
This function will add text to a \code{\link[=gg_calendar]{gg_calendar()}} plot, which can either be
the day of the month (like an ordinary calendar) or the pollutant
concentration.
}
\seealso{
Other calendarplot annotation functions: 
\code{\link{annotate_calendar_wd}()}
}
\concept{calendarplot annotation functions}
