% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_calendar.R
\name{annotate_calendar_wd}
\alias{annotate_calendar_wd}
\title{Annotate a Calendar Plot with Wind Direction Arrows}
\usage{
annotate_calendar_wd(size = 1, colour = "black", ...)
}
\arguments{
\item{size}{The maximum size of the arrows.}

\item{colour}{The colour for the arrows.}

\item{...}{Other arguments passed to \code{\link[metR:geom_arrow]{metR::geom_arrow()}}.}
}
\description{
This function will add wind direction arrows to a \code{\link[=gg_calendar]{gg_calendar()}} plot, sized
based on wind speed. Note that this function can only be used if the original
data contained numeric "ws" and "wd" columns.
}
\seealso{
Other calendarplot annotation functions: 
\code{\link{annotate_calendar_text}()}
}
\concept{calendarplot annotation functions}
