% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_roses.R
\name{annotate_rose_text}
\alias{annotate_rose_text}
\title{Annotate a wind or pollution rose with further information}
\usage{
annotate_rose_text(
  y,
  wd = 135,
  rose_angle = 30,
  size = 3,
  alpha = 1,
  fun = "label",
  ...
)
}
\arguments{
\item{y}{Y value at which to draw the annotation.}

\item{wd}{The wind direction at which to draw the annotation. Can be either
numeric, where 0 is North and 180 is South, or a character representing a
cardinal direction ("N", "NE", "E", etc.).}

\item{rose_angle}{The number passed to the \code{angle} argument of
\code{\link[=pollutionrose]{pollutionrose()}}/\code{\link[=windrose]{windrose()}}. This is required to adjust \code{wd} due
to the way \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} interacts with \code{\link[ggplot2:coord_polar]{ggplot2::coord_polar()}}.}

\item{size}{Size of the annotation.}

\item{alpha}{Alpha value of the annotation, where 1 is opaque and 0 is
transparent.}

\item{fun}{One of "label" or "text", which will use \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}
or \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}, respectively.}

\item{...}{Other arguments to pass to \code{fun}.}
}
\description{
Add an annotation to a \code{\link[=pollutionrose]{pollutionrose()}} or \code{\link[=windrose]{windrose()}} plot. This
annotation typically shows the mean pollutant or wind speed value, along with
the percentage calm value.
}
\seealso{
Other polar annotation functions: 
\code{\link{annotate_polar_axis}()},
\code{\link{annotate_polar_wedge}()}
}
\concept{polar annotation functions}
