% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_utils.R
\name{cut_date}
\alias{cut_date}
\title{Discretise date-times into categories}
\usage{
cut_date(x, type, hemisphere = "northern")
}
\arguments{
\item{x}{date(time) vector}

\item{type}{The feature to extract from the vector. Can be any one of "year",
"month", "hour", "yearmon", "weekday", "weekend", "season", or
"seasonyear".}

\item{hemisphere}{Used to split data into seasons. Can be "northern" (the
default) or "southern".}
}
\value{
a factor vector.
}
\description{
\code{\link[=cut_date]{cut_date()}} conveniently extracts useful features from a vector of date-time
data, such as the year, season, whether or not it is a weekend, and so on.
This may be useful
}
\examples{
\dontrun{
# vector
cut_date(openair::mydata$date, "year")
# data frame
dplyr::mutate(openair::mydata, year = cut_date(year))
}
}
