% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_utils.R
\name{cut_wd}
\alias{cut_wd}
\title{Discretise wind speed into sectors}
\usage{
cut_wd(x, resolution = "medium")
}
\arguments{
\item{x}{numeric vector representing wind speed, where \code{0} is North.}

\item{resolution}{A character string representing the number of sectors to
split wind direction into. Can be `0}
}
\value{
a character vector.
}
\description{
\code{\link[=cut_wd]{cut_wd()}} cuts numeric wind speed into sectors. By default, the function
splits wind direction into 8 sectors ("N", "NE", "E", "SE", "S", "SW", "W",
"NW"), it can also be split into either 4 or 16.
}
\examples{
\dontrun{
# vector
cut_wd(openair::mydata$wd)
# data frame
dplyr::mutate(openair::mydata, wd_bin = cut_wd(wd))
}
}
