% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_calendar.R
\name{gg_calendar}
\alias{gg_calendar}
\title{Plot time series values in a conventional calendar format}
\usage{
gg_calendar(
  data,
  pollutant,
  statistic = "mean",
  data_thresh = 0,
  border_colour = "white",
  w_shift = 0
)
}
\arguments{
\item{data}{A data frame minimally containing \code{date} and at least one
other numeric variable. The date should be in either \code{Date} format or
class \code{POSIXct}.}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \code{pollutant = "nox".}}

\item{statistic}{Statistic passed to \code{\link[=time_average]{time_average()}}.}

\item{data_thresh}{Data capture threshold passed to \code{\link[=time_average]{time_average()}}. For
example, \code{data_thresh = 75} means that at least 75\\% of the data must
be available in a day for the value to be calculate, else the data is
removed.}

\item{border_colour}{The colour to use for the border of each tile. Defaults
to "white". \code{NA} removes the border.}

\item{w_shift}{Controls the order of the days of the week. By default the
plot shows Saturday first (\code{w_shift = 0}). To change this so that it
starts on a Monday for example, set \code{w_shift = 2}, and so on.}
}
\description{
With a single year of data, \code{\link[=gg_calendar]{gg_calendar()}} will plot data in a conventional
calendar format, i.e., by month and day of the week. With multiple years of
data, a year-month matrix of panels will instead be plotted. Daily statistics
are calculated using \code{\link[=time_average]{time_average()}}, which by default will calculate the
daily mean concentration.
}
\details{
\code{\link[=gg_calendar]{gg_calendar()}} has two accompanying annotation functions.
\code{\link[=annotate_calendar_text]{annotate_calendar_text()}} can write either the day of the month or the
average pollutant concentration on the calendar. \code{\link[=annotate_calendar_wd]{annotate_calendar_wd()}}
will draw wind speed and direction arrows onto the calendar, assuming columns
labelled "ws" and "wd" were present in the original data.

Note that is is possible to pre-calculate concentrations in some way before
passing the data to \code{\link[=gg_calendar]{gg_calendar()}}. For example \code{\link[openair:rollingMean]{openair::rollingMean()}}
could be used to calculate rolling 8-hour mean concentrations. The data can
then be passed to \code{\link[=gg_calendar]{gg_calendar()}} and \code{statistic = "max"} chosen, which
will plot maximum daily 8-hour mean concentrations.
}
\examples{
\dontrun{
marylebone \%>\%
  selectByDate(year = 2019) \%>\%
  gg_calendar("nox") +
  annotate_calendar_text(colour = "white", size = 5, type = "date") +
  annotate_calendar_wd(colour = "black")
}
}
\seealso{
\code{\link[=annotate_calendar_wd]{annotate_calendar_wd()}} and \code{\link[=annotate_calendar_text]{annotate_calendar_text()}} for
annotating calendar plots.
}
