% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_polar_annulus.R
\name{gg_polar_annulus}
\alias{gg_polar_annulus}
\title{Bivariate polarAnnulus plot}
\usage{
gg_polar_annulus(
  data,
  pollutant,
  local_tz = NULL,
  period = "hour",
  facet = NULL,
  statistic = "mean",
  percentile = NA,
  width = 1,
  min_bin = 1,
  exclude_missing = TRUE,
  pad_date = FALSE,
  force_positive = TRUE,
  k = c(20, 10),
  normalise = FALSE,
  alpha = 1
)
}
\arguments{
\item{data}{A data frame containing wind direction, wind speed, and pollutant
concentrations.}

\item{pollutant}{One or more column names identifying pollutant
concentrations. When multiple pollutants are specified for a
single-pollutant \code{statistic} (e.g., "mean"), a faceted plot will be
returned. Two pollutants must be provided for certain \code{statistic}
options (e.g., "Pearson" in \code{\link[=gg_polar_plot]{gg_polar_plot()}}).}

\item{local_tz}{Should the results be calculated in local time that includes
a treatment of daylight savings time (DST)? The default is not to consider
DST issues, provided the data were imported without a DST offset. Emissions
activity tends to occur at local time e.g. rush hour is at 8 am every day.
When the clocks go forward in spring, the emissions are effectively
released into the atmosphere typically 1 hour earlier during the summertime
i.e. when DST applies. When plotting diurnal profiles, this has the effect
of \dQuote{smearing-out} the concentrations. Sometimes, a useful approach
is to express time as local time. This correction tends to produce
better-defined diurnal profiles of concentration (or other variables) and
allows a better comparison to be made with emissions/activity data. If set
to \code{FALSE} then GMT is used. Examples of usage include \code{local_tz
  = "Europe/London"}, \code{local_tz = "America/New_York"}. See
\code{cutData} and \code{import} for more details.}

\item{period}{This determines the temporal period to consider. Options are
\dQuote{hour} (the default, to plot diurnal variations), \dQuote{season} to
plot variation throughout the year, \dQuote{weekday} to plot day of the
week variation and \dQuote{trend} to plot the trend by wind direction.}

\item{facet}{One or two faceting columns. \code{facet} determines how the
data are split and then plotted. When \code{facet} is length 1 it is passed
to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, and when it is length 2 it is passed to
\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} with the first element being used as columns and
the second rows. Some other options (e.g., multiple \code{pollutant}
columns) can limit the the number of faceting columns to 1.}

\item{statistic}{The statistic that should be applied to each wind
speed/direction bin. Can be \dQuote{mean} (default), \dQuote{median},
\dQuote{max} (maximum), \dQuote{frequency}. \dQuote{stdev} (standard
deviation), \dQuote{weighted.mean} or \dQuote{cpf} (Conditional Probability
Function). Because of the smoothing involved, the colour scale for some of
these statistics is only to provide an indication of overall pattern and
should not be interpreted in concentration units e.g. for \code{statistic =
  "weighted.mean"} where the bin mean is multiplied by the bin frequency and
divided by the total frequency. In many cases using \code{polarFreq} will
be better. Setting \code{statistic = "weighted.mean"} can be useful because
it provides an indication of the concentration * frequency of occurrence
and will highlight the wind speed/direction conditions that dominate the
overall mean.}

\item{percentile}{If \code{statistic = "percentile"} or \code{statistic =
  "cpf"} then \code{percentile} is used, expressed from 0 to 100. Note that
the percentile value is calculated in the wind speed, wind direction
\sQuote{bins}. For this reason it can also be useful to set \code{min_bin}
to ensure there are a sufficient number of points available to estimate a
percentile. See \code{quantile} for more details of how percentiles are
calculated.}

\item{width}{The relative width of the annulus compared to the width of the
inner white space. \code{width = 2} makes the annulus twice as wide as the inner
circle, whereas \code{width = 0.5} makes the annulus half as wide as the inner
circle. Defaults to \code{1}.}

\item{min_bin}{The minimum number of points allowed in a wind speed/wind
direction bin.  The default is 1. A value of two requires at least 2 valid
records in each bin an so on; bins with less than 2 valid records are set
to NA. Care should be taken when using a value > 1 because of the risk of
removing real data points.}

\item{exclude_missing}{Setting this option to \code{TRUE} (the default)
removes points from the plot that are too far from the original data. The
smoothing routines will produce predictions at points where no data exist
i.e. they predict. By removing the points too far from the original data
produces a plot where it is clear where the original data lie. If set to
\code{FALSE} missing data will be interpolated.}

\item{pad_date}{For \code{type = "trend"} (default), \code{pad_date = TRUE}
will pad-out missing data to the beginning of the first year and the end of
the last year. The purpose is to ensure that the trend plot begins and ends
at the beginning or end of year.}

\item{force_positive}{The default is \code{TRUE}. Sometimes if smoothing data
with steep gradients it is possible for predicted values to be negative.
\code{force_positive = TRUE} ensures that predictions remain positive. This
is useful for several reasons. First, with lots of missing data more
interpolation is needed and this can result in artifacts because the
predictions are too far from the original data. Second, if it is known
beforehand that the data are all positive, then this option carries that
assumption through to the prediction. The only likely time where setting
\code{force_positive = FALSE} would be if background concentrations were
first subtracted resulting in data that is legitimately negative. For the
vast majority of situations it is expected that the user will not need to
alter the default option.}

\item{k}{The smoothing value supplied to \code{gam} for the temporal and wind
direction components, respectively. In some cases e.g. a trend plot with
less than 1-year of data the smoothing with the default values may become
too noisy and affected more by outliers. Choosing a lower value of \code{k}
(say 10) may help produce a better plot.}

\item{normalise}{If \code{TRUE} concentrations are normalised by dividing by
their mean value. This is done \emph{after} fitting the smooth surface.
This option is particularly useful if one is interested in the patterns of
concentrations for several pollutants on different scales e.g. NOx and CO.
Often useful if more than one \code{pollutant} is chosen.}

\item{alpha}{The transparency of the plot. This is mainly useful to overlay
it on a map.}
}
\value{
As well as generating the plot itself, \code{polarAnnulus} also
returns an object of class ``openair''. The object includes three main
components: \code{call}, the command used to generate the plot;
\code{data}, the data frame of summarised information used to make the
plot; and \code{plot}, the plot itself. If retained, e.g. using
\code{output <- polarAnnulus(mydata, "nox")}, this output can be used to
recover the data, reproduce or rework the original plot or undertake
further analysis.

An openair output can be manipulated using a number of generic operations,
including \code{print}, \code{plot} and \code{summary}.
}
\description{
Typically plots the concentration of a pollutant by wind direction and as a
function of time as an annulus. The function is good for visualising how
concentrations of pollutants vary by wind direction and a time period, e.g.,
by month, day of week, etc.
}
\details{
\code{\link[=gg_polar_annulus]{gg_polar_annulus()}} shares many of the properties of the \code{\link[=gg_polar_plot]{gg_polar_plot()}}.
However, \code{\link[=gg_polar_annulus]{gg_polar_annulus()}} is focussed on displaying information on how
concentrations of a pollutant (values of another variable) vary with wind
direction and time. Plotting as an annulus helps to reduce compression of
information towards the centre of the plot. The circular plot is easy to
interpret because wind direction is most easily understood in polar rather
than Cartesian coordinates.

The inner part of the annulus represents the earliest time and the outer part
of the annulus the latest time. The time dimension can be shown in many ways
including "trend", "hour" (hour or day), "season" (month of the year) and
"weekday" (day of the week). Taking hour as an example, the plot will show
how concentrations vary by hour of the day and wind direction. Such plots can
be very useful for understanding how different source influences affect a
location.

For \code{type = "trend"} the amount of smoothing does not vary linearly with
the length of the time series, i.e., a certain amount of smoothing per unit
interval in time. This is a deliberate choice because should one be
interested in a subset (in time) of data, more detail will be provided for
the subset compared with the full data set. This allows users to investigate
specific periods in more detail. Full flexibility is given through the
smoothing parameter \code{k}.
}
\seealso{
Other polar directional analysis plotting functions: 
\code{\link{gg_biasrose}()},
\code{\link{gg_polar_freq}()},
\code{\link{gg_polar_percentile}()},
\code{\link{gg_polar_plot}()},
\code{\link{gg_pollutionrose}()},
\code{\link{gg_windrose}()}
}
\concept{polar directional analysis plotting functions}
