% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_polar_percentile.R
\name{gg_polar_percentile}
\alias{gg_polar_percentile}
\title{Function to plot percentiles by wind direction}
\usage{
gg_polar_percentile(
  data,
  pollutant,
  facet = NULL,
  percentile = c(25, 50, 75, 90, 95),
  method = "default",
  line_lty = 1,
  line_width = 1,
  mean = TRUE,
  mean_lty = 1,
  mean_width = 1,
  mean_colour = "grey",
  alpha = 1
)
}
\arguments{
\item{data}{A data frame containing wind direction, wind speed, and pollutant
concentrations.}

\item{pollutant}{One or more column names identifying pollutant
concentrations. When multiple pollutants are specified for a
single-pollutant \code{statistic} (e.g., "mean"), a faceted plot will be
returned. Two pollutants must be provided for certain \code{statistic}
options (e.g., "Pearson" in \code{\link[=gg_polar_plot]{gg_polar_plot()}}).}

\item{facet}{One or two faceting columns. \code{facet} determines how the
data are split and then plotted. When \code{facet} is length 1 it is passed
to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, and when it is length 2 it is passed to
\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} with the first element being used as columns and
the second rows. Some other options (e.g., multiple \code{pollutant}
columns) can limit the the number of faceting columns to 1.}

\item{percentile}{The percentile value(s) to plot. Must be between
0--100. If \code{percentile = NA} then only a mean line will be
shown.}

\item{method}{When \code{method = "default"} the supplied
percentiles by wind direction are calculated. When \code{method
  = "cpf"} the conditional probability function (CPF) is plotted
and a single (usually high) percentile level is supplied. The
CPF is defined as CPF = my/ny, where my is the number of samples
in the wind sector y with mixing ratios greater than the
\emph{overall} percentile concentration, and ny is the total
number of samples in the same wind sector (see Ashbaugh et al.,
1985).}

\item{line_lty}{Line type for the percentile lines (see \link[ggplot2:aes_linetype_size_shape]{ggplot2::linetype}).}

\item{line_width}{Line width for the percentile lines.}

\item{mean}{Show the mean by wind direction as a line?}

\item{mean_lty}{Line type for mean line (see \link[ggplot2:aes_linetype_size_shape]{ggplot2::linetype}).}

\item{mean_width}{Line width for mean line.}

\item{mean_colour}{Line colour for mean line.}

\item{alpha}{The transparency of the plot. This is mainly useful to overlay
it on a map.}
}
\description{
Function to plot percentiles by wind direction
}
\seealso{
Other polar directional analysis plotting functions: 
\code{\link{gg_polar_annulus}()},
\code{\link{gg_polar_freq}()},
\code{\link{gg_polar_plot}()},
\code{\link{gg_pollutionrose}()},
\code{\link{gg_windrose}()}
}
\concept{polar directional analysis plotting functions}
