% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_traj_plot.R
\name{gg_traj_level}
\alias{gg_traj_level}
\title{Trajectory level plots with conditioning}
\usage{
gg_traj_level(
  data,
  lon = "lon",
  lat = "lat",
  statistic = "frequency",
  pollutant = "height",
  lat_inc = 1,
  lon_inc = 1,
  min_bin = 1,
  percentile = 90,
  sqtba_combine = NA,
  sqtba_sigma = 1.5,
  border_colour = NA,
  map = TRUE,
  map_fill = "grey95",
  map_colour = "grey85",
  map_alpha = 0.5,
  projection = "mercator",
  ...
)
}
\arguments{
\item{data}{Data frame, the result of importing a trajectory file using
\code{\link[openair:importTraj]{openair::importTraj()}}.}

\item{lon}{Column containing the longitude, as a decimal.}

\item{lat}{Column containing the latitude, as a decimal.}

\item{statistic}{By default the function will plot the trajectory
frequencies. There are also various ways of plotting concentrations.

It is possible to set \code{statistic = "difference"}. In this case
trajectories where the associated concentration is greater than
\code{percentile} are compared with the the full set of trajectories to
understand the differences in frequencies of the origin of air masses. The
comparison is made by comparing the percentage change in gridded
frequencies. For example, such a plot could show that the top 10\\% of
concentrations of PM10 tend to originate from air-mass origins to the east.

If \code{statistic = "pscf"} then a Potential Source Contribution Function
map is produced. If \code{statistic = "cwt"} then concentration weighted
trajectories are plotted.

If \code{statistic = "cwt"} then the Concentration Weighted Trajectory
approach is used. See details.}

\item{pollutant}{By default, the trajectory height is summarised. \code{pollutant}
overrides this, which is useful when \code{statistic} is something other than
\code{"frequency"}.}

\item{lat_inc, lon_inc}{The latitude- and longitude-interval to be used for
binning data.}

\item{min_bin}{The minimum number of unique points in a grid cell. Counts
below \code{min_bin} are set as missing.}

\item{percentile}{For \code{trajLevel}. The percentile concentration of
\code{pollutant} against which the all trajectories are compared.}

\item{sqtba_combine}{When statistic is "SQTBA" it is possible to combine lots
of receptor locations to derive a single map. \code{.combine} identifies
the column that differentiates different sites (commonly a column named
\code{site}). Note that individual site maps are normalised first by
dividing by their mean value.}

\item{sqtba_sigma}{For the SQTBA approach \code{sigma} determines the amount
of back trajectory spread based on the Gaussian plume equation. Values in
the literature suggest 5.4 km after one hour. However, testing suggests
lower values reveal source regions more effectively while not introducing
too much noise.}

\item{border_colour}{The colour to use for the border of each tile. Defaults
to \code{NA}, which removes the border.}

\item{map}{Should a base map be drawn? If \code{TRUE} the world base map from
the \code{maps} package is used.}

\item{map_fill}{Fill colour of map.}

\item{map_colour}{Line colour of map.}

\item{map_alpha}{Alpha of map.}

\item{projection}{The map projection to be used. Different map projections
are possible through the \code{mapproj} package. See \code{?mapproject} for
extensive details and information on setting other parameters and
orientation (see below).}
}
\description{
This function plots grided back trajectories. This function requires that
data are imported using the \code{\link[openair:importTraj]{openair::importTraj()}} function.
}
\seealso{
Other trajectory analysis functions: 
\code{\link{gg_traj_plot}()},
\code{\link{traj_cluster}()}
}
\concept{trajectory analysis functions}
