% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_traj_plot.R
\name{gg_traj_plot}
\alias{gg_traj_plot}
\title{Trajectory line plots with conditioning}
\usage{
gg_traj_plot(
  data,
  lon = "lon",
  lat = "lat",
  colour = NULL,
  facet = NULL,
  alpha = 1,
  n_points = 12,
  map = TRUE,
  map_fill = "grey95",
  map_colour = "grey85",
  map_alpha = 0.5,
  projection = "mercator",
  ...
)
}
\arguments{
\item{data}{Data frame, the result of importing a trajectory file using
\code{\link[openair:importTraj]{openair::importTraj()}}.}

\item{lon}{Column containing the longitude, as a decimal.}

\item{lat}{Column containing the latitude, as a decimal.}

\item{colour}{A column to colour the paths by.}

\item{facet}{How to facet the plot.}

\item{alpha}{Transparency of the plots/lines, where \code{1} is completely opaque
and \code{0} is completely transparent.}

\item{n_points}{A dot is placed every \code{n_points} along each full
trajectory. For hourly back trajectories points are plotted every
\code{npoint} hours. This helps to understand where the air masses were at
particular times and get a feel for the speed of the air (points closer
together correspond to slower moving air masses). If \code{n_points = NULL}
then no points are added.}

\item{map}{Should a base map be drawn? If \code{TRUE} the world base map from
the \code{maps} package is used.}

\item{map_fill}{Fill colour of map.}

\item{map_colour}{Line colour of map.}

\item{map_alpha}{Alpha of map.}

\item{projection}{The map projection to be used. Different map projections
are possible through the \code{mapproj} package. See \code{?mapproject} for
extensive details and information on setting other parameters and
orientation (see below).}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:coord_map]{ggplot2::coord_map}}
  \describe{
    \item{\code{orientation}}{projection orientation, which defaults to
\code{c(90, 0, mean(range(x)))}.  This is not optimal for many
projections, so you will have to supply your own. See
\code{\link[mapproj:mapproject]{mapproj::mapproject()}} for more information.}
    \item{\code{clip}}{Should drawing be clipped to the extent of the plot panel? A
setting of \code{"on"} (the default) means yes, and a setting of \code{"off"}
means no. For details, please see \code{\link[ggplot2:coord_cartesian]{coord_cartesian()}}.}
  }}
}
\description{
This function plots back trajectories. This function requires that data are
imported using the \code{\link[openair:importTraj]{openair::importTraj()}} function.
}
