% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_polar_roses.R
\name{gg_windrose}
\alias{gg_windrose}
\title{Traditional wind rose plot and pollution rose variation}
\usage{
gg_windrose(
  data,
  ws = "ws",
  wd = "wd",
  ws_int = 2,
  angle = 30,
  facet = NULL,
  bias_corr = TRUE,
  breaks = 4,
  normalise = FALSE,
  statistic = "prop.count",
  width = 0.9,
  border_colour = NA,
  alpha = 1
)
}
\arguments{
\item{data}{A data frame containing fields \code{ws} and \code{wd}}

\item{ws}{Name of the column representing wind speed.}

\item{wd}{Name of the column representing wind direction.}

\item{ws_int}{The Wind speed interval. Default is 2 m/s but for low met masts
with low mean wind speeds a value of 1 or 0.5 m/s may be better. Note, this
argument is superseded in \code{pollutionRose}. See \code{breaks} below.}

\item{angle}{Angle of the spokes. Ideally a number by which 360 is evenly
divisible (see \code{bias_corr}).}

\item{facet}{One or two faceting columns. \code{facet} determines how the
data are split and then plotted. When \code{facet} is length 1 it is passed
to [ggplot2::facet_wrap()], and when it is length 2 it is passed to
[ggplot2::facet_grid()] with the first element being used as columns and
the second rows. Some other options (e.g., multiple \code{pollutant}
columns) can limit the the number of faceting columns to 1.}

\item{bias_corr}{When \code{angle} does not divide exactly into 360 a bias is
introduced in the frequencies when the wind direction is already supplied
rounded to the nearest 10 degrees, as is often the case. For example, if
\code{angle = 22.5}, N, E, S, W will include 3 wind sectors and all other
angles will be two. A bias correction can made to correct for this problem.
A simple method according to Applequist (2012) is used to adjust the
frequencies.}

\item{breaks}{The number of break points to use when binning the data (used
alongside \code{ws_int}), or a vector of specific break points. For
example, \code{breaks = 6} and \code{ws_int = 2} generates the break points
2, 4, 6 & 8 m/s, whereas \code{breaks = c(0, 1, 10, 100)} breaks the data
into four segments: <1, 1-10, 10-100, & >100.}

\item{normalise}{If \code{TRUE} each wind direction segment is normalised to
equal one. This is useful for showing how the concentrations (or other
parameters) contribute to each wind sector when the proprtion of time the
wind is from that direction is low. A line showing the probability that the
wind directions is from a particular wind sector is also shown.}

\item{statistic}{The \code{statistic} to be applied to each data bin in the
plot. Options currently include \dQuote{prop.count}, \dQuote{prop.mean} and
\dQuote{abs.count}. The default \dQuote{prop.count} sizes bins according to
the proportion of the frequency of measurements.  Similarly,
\dQuote{prop.mean} sizes bins according to their relative contribution to
the mean. \dQuote{abs.count} provides the absolute count of measurements in
each bin.}

\item{width}{Width of each bar as a fraction of its maximum width.
\code{width = 1} makes all bars meet at their edges and \code{width = 0}
makes them disappear entirely. Defaults to \code{0.9}.}

\item{border_colour}{Border colour for shaded areas. Default is no border.}

\item{alpha}{The transparency of the plot. This is mainly useful to overlay
it on a map.}
}
\description{
Traditional wind rose plot and pollution rose variation
}
\seealso{
Other polar directional analysis plotting functions: 
\code{\link{gg_polar_annulus}()},
\code{\link{gg_polar_freq}()},
\code{\link{gg_polar_percentile}()},
\code{\link{gg_polar_plot}()},
\code{\link{gg_pollutionrose}()}
}
\concept{polar directional analysis plotting functions}
