% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggopenair-package.R
\docType{package}
\name{ggopenair-package}
\alias{ggopenair}
\alias{ggopenair-package}
\title{ggopenair: Air Pollution Data Visualisation using ggplot2}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Recreate the air quality data visualisations of 'openair' using the popular 'ggplot2' plotting package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jack-davison.github.io/ggopenair/}
}

}
\author{
\strong{Maintainer}: Jack Davison \email{davison.jack.jd@gmail.com}

Authors:
\itemize{
  \item David Carslaw \email{david.carslaw@ricardo.com}
}

}
\keyword{internal}
