% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_freq.R
\name{polar_freq}
\alias{polar_freq}
\title{Function to plot wind speed/direction frequencies and other statistics}
\usage{
polar_freq(
  data,
  pollutant = NULL,
  statistic = "frequency",
  ws_int = 1,
  wd_nint = 36,
  facet = NULL,
  min_bin = 1,
  border_colour = NA,
  alpha = 1
)
}
\arguments{
\item{data}{A data frame containing wind direction, wind speed, and pollutant
concentrations.}

\item{pollutant}{A column name identifying pollutant concentration.}

\item{statistic}{The statistic that should be applied to each wind
speed/direction bin. Can be \dQuote{frequency}, \dQuote{mean},
\dQuote{median}, \dQuote{max} (maximum), \dQuote{stdev} (standard
deviation) or \dQuote{weighted.mean}. The option \dQuote{frequency} (the
default) is the simplest and plots the frequency of wind speed/direction in
different bins. The scale therefore shows the counts in each bin. The
option \dQuote{mean} will plot the mean concentration of a pollutant (see
next point) in wind speed/direction bins, and so on.  Finally,
\dQuote{weighted.mean} will plot the concentration of a pollutant weighted
by wind speed/direction. Each segment therefore provides the percentage
overall contribution to the total concentration. More information is given
in the examples. Note that for options other than \dQuote{frequency}, it is
necessary to also provide the name of a pollutant. See function
\code{cutData} for further details.}

\item{ws_int}{Wind speed interval assumed. In some cases e.g. a low met mast,
an interval of 0.5 may be more appropriate.}

\item{wd_nint}{Number of intervals of wind direction.}

\item{facet}{One or two faceting columns. \code{facet} determines how the
data are split and then plotted. When \code{facet} is length 1 it is passed
to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, and when it is length 2 it is passed to
\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} with the first element being used as columns and
the second rows. Some other options (e.g., multiple \code{pollutant}
columns) can limit the the number of faceting columns to 1.}

\item{min_bin}{The minimum number of points allowed in a wind speed/wind
direction bin.  The default is 1. A value of two requires at least 2 valid
records in each bin an so on; bins with less than 2 valid records are set
to NA. Care should be taken when using a value > 1 because of the risk of
removing real data points.}

\item{border_colour}{The colour to use for the border of each tile. Defaults
to \code{NA}, which removes the border.}

\item{alpha}{The transparency of the plot. This is mainly useful to overlay
it on a map.}
}
\description{
Function to plot wind speed/direction frequencies and other statistics
}
\seealso{
Other polar directional analysis functions: 
\code{\link{polar_annulus}()},
\code{\link{polar_cluster}()},
\code{\link{polar_diff}()},
\code{\link{polar_plot}()},
\code{\link{pollutionrose}()},
\code{\link{rose_metbias}()},
\code{\link{rose_percentile}()},
\code{\link{rose_wind}()}
}
\concept{polar directional analysis functions}
