% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_text.R
\name{quick_text}
\alias{quick_text}
\title{Automatic text formatting for ggopenair}
\usage{
quick_text(text)
}
\arguments{
\item{text}{A character vector.}
}
\value{
A character vector.
}
\description{
[quick_text()] is routine formatting lookup table which automatically
prepares common pollutant names and units for parsing using
[ggplot2::label_parsed()] or similar. It is automatically called by several
\code{ggopenair} functions. Unlike [openair::quickText()], [quick_text()]
returns a character vector, which means its output can be stored in a data
frame or [tibble::tibble()]
}
