% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_roses.R
\name{rose_metbias}
\alias{rose_metbias}
\title{Bias Rose}
\usage{
rose_metbias(
  data,
  ws = "ws",
  wd = "wd",
  ws2 = "ws2",
  wd2 = "wd2",
  angle = 10,
  facet = NULL,
  bias_corr = TRUE,
  breaks = NA,
  width = 1,
  border_colour = NA,
  alpha = 1
)
}
\arguments{
\item{data}{A data frame containing fields \code{ws}, \code{wd}, \code{ws2}
and \code{wd2}.}

\item{ws}{Name of the column representing wind speed.}

\item{wd}{Name of the column representing wind direction.}

\item{ws2}{Name of the column representing wind speed (2).}

\item{wd2}{Name of the column representing wind direction (2).}

\item{angle}{Angle of the spokes. Ideally a number by which 360 is evenly
divisible (see \code{bias_corr}).}

\item{facet}{One or two faceting columns. \code{facet} determines how the
data are split and then plotted. When \code{facet} is length 1 it is passed
to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, and when it is length 2 it is passed to
\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} with the first element being used as columns and
the second rows. Some other options (e.g., multiple \code{pollutant}
columns) can limit the the number of faceting columns to 1.}

\item{bias_corr}{When \code{angle} does not divide exactly into 360 a bias is
introduced in the frequencies when the wind direction is already supplied
rounded to the nearest 10 degrees, as is often the case. For example, if
\code{angle = 22.5}, N, E, S, W will include 3 wind sectors and all other
angles will be two. A bias correction can made to correct for this problem.
A simple method according to Applequist (2012) is used to adjust the
frequencies.}

\item{breaks}{The number of break points to use when binning the data (used
alongside \code{ws_int}), or a vector of specific break points. For
example, \code{breaks = 6} and \code{ws_int = 2} generates the break points
2, 4, 6 & 8 m/s, whereas \code{breaks = c(0, 1, 10, 100)} breaks the data
into four segments: <1, 1-10, 10-100, & >100.}

\item{width}{Width of each bar as a fraction of its maximum width.
\code{width = 1} makes all bars meet at their edges and \code{width = 0}
makes them disappear entirely. Defaults to \code{0.9}.}

\item{border_colour}{Border colour for shaded areas. Default is no border.}

\item{alpha}{The transparency of the plot. This is mainly useful to overlay
it on a map.}
}
\description{
Bias Rose
}
\seealso{
Other polar directional analysis functions: 
\code{\link{polar_annulus}()},
\code{\link{polar_cluster}()},
\code{\link{polar_diff}()},
\code{\link{polar_freq}()},
\code{\link{polar_plot}()},
\code{\link{pollutionrose}()},
\code{\link{rose_percentile}()},
\code{\link{rose_wind}()}
}
\concept{polar directional analysis functions}
