% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_limitval.R
\name{scale_limitval}
\alias{scale_y_limitval}
\alias{scale_x_limitval}
\title{Position scales for continuous data (x & y) with labelled vertical/horizontal
markers}
\usage{
scale_y_limitval(
  marker_values,
  marker_colours = "black",
  marker_labels = marker_values,
  trans = "identity",
  ...
)

scale_x_limitval(
  marker_values,
  marker_colours = "black",
  marker_labels = marker_values,
  trans = "identity",
  ...
)
}
\arguments{
\item{marker_values}{Numeric vector of values at which to draw marker lines.}

\item{marker_colours}{Character vector of colours for marker lines. Should be
the same length as \code{marker_values}.}

\item{marker_labels}{Character vector of labels for marker lines. Should be
the same length as \code{marker_values}. Defaults to using the numeric
values given in \code{marker_values}.}

\item{trans}{The name of a transformation object. See
[ggplot2::scale_continuous()] for more information.}

\item{...}{Other arguments to pass to [ggplot2::scale_continuous()].}
}
\description{
This function is a wrapper around [ggplot2::scale_continuous()] wich allows
users to automatically draw labelled horizontal or vertical markers at
specified points. In the air quality world, this may be particularly useful
to display limit values, such as those defined in legislation or recommended
by health authorities.
}
\details{
This function uses [ggplot2::sec_axis()] to display the line labels. One can
therefore not pass [ggplot2::sec_axis()] to \code{...}. If users wish to use
[ggplot2::sec_axis()] it is recommended to use [ggplot2::geom_abline()] to
draw lines manually.

Unlike other "scales" functions, including [ggplot2::scale_continuous()], the
order in which this function is added to the [ggplot2::ggplot()] object
matters. Adding \code{scale_*_limitval()} *after* [ggplot2::geom_line()] will
draw the line markers on top of the trend line, whereas adding it *before*
will draw the markers below.
}
