% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_polar_utils.R
\name{annotate_polar_axis}
\alias{annotate_polar_axis}
\title{Annotate a Polar Plot with Axis Labels}
\usage{
annotate_polar_axis(
  breaks,
  direction = "NW",
  drop = TRUE,
  ...,
  .rose_angle = NULL
)
}
\arguments{
\item{breaks}{The axis breaks to label on the plot. It may be appropriate to
use \code{\link[=seq]{seq()}} to obtain equally spaced labels.}

\item{direction}{The angles at which to write the labels. Can be expressed
numerically, or using cardinal direction abbreviations ("N", "NNE", "NE",
"ENE", etc.).}

\item{drop}{Remove the y-axis labels to the side of the plot? Defaults to
\code{TRUE}. Note that this option is overriden by complete themes such as
\code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}} or \code{\link[=theme_polar]{theme_polar()}}.}

\item{...}{Arguments to pass to \code{\link[ggplot2:annotate]{ggplot2::annotate()}}.}

\item{.rose_angle}{The number passed to the \code{angle} argument of
\code{\link[=rose_pollution]{rose_pollution()}}/\code{\link[=rose_wind]{rose_wind()}}. This is required to adjust \code{wd} when
it is near to North/0 due to the way \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} interacts with
\code{\link[ggplot2:coord_polar]{ggplot2::coord_polar()}}.}
}
\value{
An annotation to be added to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
This is a wrapper around \code{\link[ggplot2:annotate]{ggplot2::annotate()}} which draws the y-axis labels
directly on the plot, and optionally drops the labels at the side. Note that,
owing to the way \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} interacts with
\code{\link[ggplot2:coord_polar]{ggplot2::coord_polar()}}, the \code{.rose_angle} argument should be used with
\code{\link[=rose_pollution]{rose_pollution()}} and \code{\link[=rose_wind]{rose_wind()}}.
}
\examples{
\dontrun{
polar_plot(mydata, "nox") + annotate_polar_axis(seq(5, 25, 5))
}

}
\seealso{
Other polar annotation functions: 
\code{\link{annotate_polar_wedge}()},
\code{\link{annotate_rose_text}()}
}
\concept{polar annotation functions}
