% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_polar_utils.R
\name{annotate_polar_wedge}
\alias{annotate_polar_wedge}
\title{Annotate a Polar Plot with a Coloured Wedge}
\usage{
annotate_polar_wedge(
  start,
  end,
  fill = "red",
  colour = NA,
  alpha = 0.25,
  ...,
  .rose_angle = NULL
)
}
\arguments{
\item{start}{The angle at which to start the wedge. Can be expressed
numerically, or using cardinal direction abbreviations ("N", "NNE", "NE",
"ENE", etc.).}

\item{end}{The angle at which to end the wedge. See \code{end} for more details.}

\item{fill}{The fill colour of the wedge. Defaults to "red".}

\item{colour}{The border colour of the wedge. Defaults to \code{NA}, which means
no border is drawn.}

\item{alpha}{The transparency of the wedge. \code{1} is totally opaque and \code{0} is
completely transparent.}

\item{...}{Arguments to pass to \code{\link[ggplot2:annotate]{ggplot2::annotate()}}.}

\item{.rose_angle}{The number passed to the \code{angle} argument of
\code{\link[=gg_pollutionrose]{gg_pollutionrose()}}/\code{\link[=gg_windrose]{gg_windrose()}}. This is required to adjust \code{start}
and \code{end} when they are near to North/0 due to the way
\code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} interacts with \code{\link[ggplot2:coord_polar]{ggplot2::coord_polar()}}.}
}
\value{
An annotation to be added to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
This is a wrapper around \code{\link[ggplot2:annotate]{ggplot2::annotate()}} which draws a semi-transparent
wedge on a plot with continuous polar coordinates (any of \code{\link[=gg_polar_plot]{gg_polar_plot()}},
\code{\link[=gg_polar_annulus]{gg_polar_annulus()}}, \code{\link[=gg_polar_freq]{gg_polar_freq()}} or \code{\link[=gg_polar_percentile]{gg_polar_percentile()}}).
}
\examples{
\dontrun{
gg_polar_plot(mydata, "nox") + annotate_polar_wedge(start = "N", end = "E")
}

}
\seealso{
Other polar annotation functions: 
\code{\link{annotate_polar_axis}()},
\code{\link{annotate_rose_text}()}
}
\concept{polar annotation functions}
