% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_polar_utils.R
\name{annotate_polar_wedge}
\alias{annotate_polar_wedge}
\title{Annotate a Polar Plot with a Coloured Wedge}
\usage{
annotate_polar_wedge(
  start,
  end,
  fill = "red",
  colour = NA,
  alpha = 0.25,
  ...,
  .rose_angle = NULL
)
}
\arguments{
\item{start}{The angle at which to start the wedge. Can be expressed
numerically, or using cardinal direction abbreviations ("N", "NNE", "NE",
"ENE", etc.).}

\item{end}{The angle at which to end the wedge. See \code{end} for more details.}

\item{fill}{The fill colour of the wedge. Defaults to "red".}

\item{colour}{The border colour of the wedge. Defaults to \code{NA}, which means
no border is drawn.}

\item{alpha}{The transparency of the wedge. \code{1} is totally opaque and \code{0} is
completely transparent.}

\item{...}{Arguments to pass to \code{\link[ggplot2:annotate]{ggplot2::annotate()}}.}

\item{.rose_angle}{The number passed to the \code{angle} argument of
\code{\link[=pollutionrose]{pollutionrose()}}/\code{\link[=windrose]{windrose()}}. This is required to adjust \code{start}
and \code{end} when they are near to North/0 due to the way
\code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} interacts with \code{\link[ggplot2:coord_polar]{ggplot2::coord_polar()}}.}
}
\value{
An annotation to be added to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
This is a wrapper around \code{\link[ggplot2:annotate]{ggplot2::annotate()}} which draws a semi-transparent
wedge on a plot with continuous polar coordinates (any of \code{\link[=polar_plot]{polar_plot()}},
\code{\link[=polar_annulus]{polar_annulus()}}, \code{\link[=polar_freq]{polar_freq()}} or \code{\link[=polar_percentile]{polar_percentile()}}). Note
that, owing to the way \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} interacts with
\code{\link[ggplot2:coord_polar]{ggplot2::coord_polar()}}, the \code{.rose_angle} argument should be used with
\code{\link[=pollutionrose]{pollutionrose()}} and \code{\link[=windrose]{windrose()}}.
}
\examples{
\dontrun{
polar_plot(mydata, "nox") + annotate_polar_wedge(start = "N", end = "E")
}

}
\seealso{
Other polar annotation functions: 
\code{\link{annotate_polar_axis}()},
\code{\link{annotate_rose_text}()}
}
\concept{polar annotation functions}
