% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_calendar.R
\name{gg_calendar}
\alias{gg_calendar}
\title{Plot time series values in convential calendar format}
\usage{
gg_calendar(
  data,
  pollutant,
  statistic = "mean",
  data_thresh = 0,
  border_colour = "white",
  w_shift = 0
)
}
\arguments{
\item{data}{A data frame minimally containing \code{date} and at least one
other numeric variable. The date should be in either \code{Date} format or
class \code{POSIXct}.}

\item{pollutant}{Mandatory. A pollutant name corresponding to a variable in a
data frame should be supplied e.g. \code{pollutant = "nox".}}

\item{statistic}{Statistic passed to [openair::timeAverage()].}

\item{data_thresh}{Data capture threshold passed to [openair::timeAverage()].
For example, \code{data_thresh = 75} means that at least 75\% of the data
must be available in a day for the value to be calculate, else the data is
removed.}

\item{border_colour}{The colour to use for the border of each tile. Defaults
to "white". `NA` removes the border.}

\item{w_shift}{Controls the order of the days of the week. By default the
plot shows Saturday first (\code{w_shift = 0}). To change this so that it
starts on a Monday for example, set \code{w_shift = 2}, and so on.}
}
\description{
Plot time series values in convential calendar format
}
\examples{
\dontrun{
marylebone \%>\%
  selectByDate(year = 2019) \%>\%
  gg_calendar("nox") +
  annotate_calendar_text(colour = "white", size = 5, type = "date") +
  annotate_calendar_wd(colour = "black")
}
}
