% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggopenair-package.R
\docType{package}
\name{ggopenair-package}
\alias{ggopenair}
\alias{ggopenair-package}
\title{ggopenair: Air Pollution Data Visualisation using ggplot2}
\description{
Recreate the air quality data visualisations of 'openair' using the popular 'ggplot2' plotting package.
}
\author{
\strong{Maintainer}: Jack Davison \email{davison.jack.jd@gmail.com}

}
\keyword{internal}
