% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_opencolours.R
\name{scale_opencolours_c}
\alias{scale_opencolours_c}
\title{Plotting with openair palettes in ggplot2}
\usage{
scale_opencolours_c(
  colours = "default",
  na_value = "grey50",
  aes = c("colour", "fill"),
  ...
)
}
\arguments{
\item{colours}{The openair colour scheme to use. See [openair::openColours()]
for more information.}

\item{na_value}{The aesthetic value to use for missing (NA) values. Defaults
to "grey50".}

\item{aes}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. The default,
`c("colour", "fill")`, applies colours to both the `colour` *and* `fill`
aesthetics.}

\item{...}{Other arguments to pass to [ggplot2::scale_color_gradientn()].}
}
\value{
A ggproto object defining a continuous colour scale for use with
  [ggplot2::ggplot()].
}
\description{
This is a convenient wrapper around [ggplot2::scale_color_gradientn()] and
[openair::openColours()] which creates colour gradients for `ggopenair`
plots, with helpful defaults.
}
\examples{
\dontrun{
gg_polar_plot(mydata, "nox") + scale_opencolours()
}

}
