% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_polar.R
\name{theme_polar}
\alias{theme_polar}
\title{Default Polar Plot Theme}
\usage{
theme_polar(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  base_panel_ontop = TRUE
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{panel_ontop}{Place panel grids on top of the plot geometries? The
default, \code{TRUE}, can make polar plots easier to interpret, but this
behaviour may not always be desired.}
}
\value{
A ggplot theme object defining for use with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
This is a complete theme which attempts to recreate the styling of
\code{\link[openair:polarPlot]{openair::polarPlot()}}. As an alternative to using this theme, you could use
\code{\link[ggplot2:theme]{ggplot2::theme()}} or any of the in-built \code{ggplot2} themes.
}
