% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj.R
\name{traj_cluster}
\alias{traj_cluster}
\title{Calculate clusters for back trajectories}
\usage{
traj_cluster(
  data,
  method = "Euclid",
  n_cluster = 5,
  facet = NULL,
  split_after = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame, the result of importing a trajectory file using
\code{\link[openair:importTraj]{openair::importTraj()}}.}

\item{method}{Method used to calculate the distance matrix for the back
trajectories. There are two methods available: "Euclid" and "Angle".}

\item{n_cluster}{Number of clusters to calculate.}

\item{facet}{One or two faceting columns. \code{facet} determines how the
data are split and then clustered. Note that the cluster calculations are
separately made of each level of "type".}

\item{split_after}{When \code{facet} != \code{NULL}, the trajectories can
either be calculated for each facet independently or extracted after the
cluster calculations have been applied to the whole data set.}

\item{plot}{Automatically plot the clustered trajectories? Defaults to
\code{FALSE}. When \code{TRUE}, the plots can be controlled using \code{...}.
Alternatively, the output of \code{\link[=traj_cluster]{traj_cluster()}} can be manually passed to
\code{\link[=traj_plot]{traj_plot()}}.}

\item{...}{
  Arguments passed on to \code{\link[=traj_plot]{traj_plot}}
  \describe{
    \item{\code{alpha}}{Transparency of the plots/lines, where \code{1} is completely opaque
and \code{0} is completely transparent.}
    \item{\code{n_points}}{A dot is placed every \code{n_points} along each full
trajectory. For hourly back trajectories points are plotted every
\code{npoint} hours. This helps to understand where the air masses were at
particular times and get a feel for the speed of the air (points closer
together correspond to slower moving air masses). If \code{n_points = NULL}
then no points are added.}
    \item{\code{map}}{Should a base map be drawn? If \code{TRUE} the world base map from
the \code{maps} package is used.}
    \item{\code{map_fill}}{Fill colour of map.}
    \item{\code{map_colour}}{Line colour of map.}
    \item{\code{map_alpha}}{Alpha of map.}
    \item{\code{lon_range,lat_range}}{Longitude and latitude range, provided as a
numeric vector of length 2.}
    \item{\code{crs}}{The coordinate reference system (CRS) used for the base map.
Defaults to \code{4326}, which behaves similarly to Cartesian coordinates. A useful
alternative could be \code{3035} (the "lambert" projection).}
  }}
}
\description{
This function carries out cluster analysis of HYSPLIT back trajectories. The
function is specifically designed to work with the trajectories imported
using the \code{\link[openair:importTraj]{openair::importTraj()}} function, which provides pre-calculated
back trajectories at specific receptor locations. Clustered trajectories can
be visualised using \code{\link[=traj_plot]{traj_plot()}}.
}
\seealso{
Other trajectory analysis functions: 
\code{\link{traj_level}()},
\code{\link{traj_plot}()}
}
\concept{trajectory analysis functions}
